Refactor scraping package into engine subdirectories
