+++ MOTEUR/scraping/image_scraper/constants.py
