+++ MOTEUR/scraping/image_scraper/product_sentences.json
